convert data models to geojson at edges, rather than in the store
