+++ src/services/effects.ts
@@ -17,7 +17,7 @@ export function withEffects(store: Store<Actions>) {
     )
     .subscribe(async ([distribution, serviceAreas]) => {
       let points = await getRepresentativePoints(distribution, serviceAreas)
-      store.set('representativePoints')(toGeoJSON(representativePointToFeature)(points))
+      store.set('representativePoints')(points)
     })
 
   /**
@@ -27,7 +27,7 @@ export function withEffects(store: Store<Actions>) {
    */
   store.on('uploadedProviders').subscribe(async providers => {
     let result = await postProviders(providers)
-    store.set('providers')(toGeoJSON(providerToFeature)(
+    store.set('providers')(
       chain(result)
         .zip<WriteProvidersResponse | WriteProvidersRequest>(providers)
         .partition(([res]: [WriteProvidersResponse]) => isWriteProvidersSuccessResponse(res))
@@ -40,51 +40,8 @@ export function withEffects(store: Store<Actions>) {
           id: res.id
         }))
         .value()
-    ))
+    )
   })
 
   return store
-}
-
-function toGeoJSON<T>(f: (point: T) => GeoJSON.Feature<GeoJSON.GeometryObject>) {
-  return (points: T[]): GeoJSON.FeatureCollection<GeoJSON.GeometryObject> => ({
-    type: 'FeatureCollection',
-    features: points.map(f)
-  })
-}
-
-function providerToFeature(
-  point: Provider
-): GeoJSON.Feature<GeoJSON.GeometryObject> {
-  return {
-    id: point.id,
-    type: 'Feature',
-    properties: {
-      address: point.address,
-      languages: point.languages,
-      npi: point.npi,
-      specialty: point.specialty
-    },
-    geometry: {
-      type: 'Point',
-      coordinates: [point.lng, point.lat]
-    }
-  }
-}
-
-function representativePointToFeature(
-  point: RepresentativePoint
-): GeoJSON.Feature<GeoJSON.GeometryObject> {
-  return {
-    id: point.id,
-    type: 'Feature',
-    properties: {
-      population: point.population,
-      service_area_id: point.service_area_id
-    },
-    geometry: {
-      type: 'Point',
-      coordinates: [point.lng, point.lat]
-    }
-  }
 }
